update(sdk/schema): Change offset data type

+++ src/sdk/schema.h
@@ -8,7 +8,7 @@
 
 struct SchemaKey
 {
-    int16_t offset;
+    int32_t offset;
     bool networked;
 };
